requirements.txt
py-solc-x
web3
python-dotenv
eth-brownie